+++ changelog.md
@@ -17,6 +17,7 @@
 * Added support for multiple Decklink outputs in Video Capture SDK .Net and Video Edit SDK .Net  (Windows-only engine)
 * API breaking change: the frame rate changed from double to VideoFrameRate in all APIs
 * Resolved issues with reverse playback engine in Media Player SDK .Net (Windows-only engine)
+* ONVIFControl and other ONVIF-related APIs are available for all platforms
 
 ## 15.4